webprak/src/test/java/ru/msu/cmc/webprak/DAO/ClientDAOTest.java
package ru.msu.cmc.webprak.DAO;
import java.util.List;

import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.junit.jupiter.api.Assertions.assertNull;
import org.hibernate.Session;
import org.hibernate.SessionFactory;
import org.junit.jupiter.api.*;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.test.context.TestPropertySource;
import ru.msu.cmc.webprak.models.Category;
import ru.msu.cmc.webprak.models.Client;

@SpringBootTest
@TestInstance(TestInstance.Lifecycle.PER_CLASS)
@TestPropertySource(locations="classpath:application.properties")
public class ClientDAOTest {

    @Autowired
    private ClientDAO clientDAO;
    @Autowired
    private SessionFactory sessionFactory;

    @Test
    void testGetClientByName() {
        List<Client> clients = clientDAO.findAllMatching("A potato consumer");
        assertEquals(1, clients.size());
    }

}
